requirements.txt
certifi==2020.12.5
chardet==4.0.0
idna==2.10
numpy==1.20.1
opencv-python==4.5.1.48
Pillow==8.1.2
PySide2==5.15.2
requests==2.25.1
shiboken2==5.15.2
typing-extensions==3.7.4.3
urllib3==1.26.4
